fix: app code refactor

+++ src/App.tsx
@@ -3,62 +3,74 @@ import './App.css';
 import {Button} from './Components/Button/Button';
 
 const App = () => {
+    const [initValue, setInitValue] = useState(0)
     const [minValue, setMinValue] = useState(0)
     const [maxValue, setMaxValue] = useState(0)
-    const [endMinValue, setEndMinValue] = useState(0)
-    const [endMaxValue, setEndMaxValue] = useState(0)
+    const [isChecked, setIsChecked] = useState(false)
+    const [radioValue, setRadioValue] = useState('Off')
     const [changeMode, setChangeMode] = useState(false)
     const [incorrect, setIncorrect] = useState(false)
     const [day, setDay] = useState(true)
 
     useEffect(() => {
-        let item = localStorage.getItem('counterValue')
+        let itemMinValue = localStorage.getItem('counterMinValue')
         let itemMaxValue = localStorage.getItem('counterMaxValue')
-        let itemEndMinValue = localStorage.getItem('counterEndMinValue')
-        let itemEndMaxValue = localStorage.getItem('counterEndMaxValue')
-        if (item) {
-            let newValue = JSON.parse(item)
-            setMinValue(newValue)
+        let itemValue = localStorage.getItem('counterInitValue')
+        if (itemMinValue) {
+            let newValue = JSON.parse(itemMinValue)
+            setMinValue(+newValue)
         }
         if (itemMaxValue) {
             let newValue = JSON.parse(itemMaxValue)
-            setMaxValue(newValue)
+            setMaxValue(+newValue)
         }
-        if (itemEndMinValue) {
-            let newValue = JSON.parse(itemEndMinValue)
-            setEndMinValue(newValue)
+        if (itemValue) {
+            let newValue = JSON.parse(itemValue)
+            setInitValue(newValue)
         }
-        if (itemEndMaxValue) {
-            let newValue = JSON.parse(itemEndMaxValue)
-            setEndMaxValue(newValue)
-        }
-        setChangeMode(true)
-    }, [])
+        setChangeMode(false)
+    }, []) //1 time
 
     useEffect(() => {
-        localStorage.setItem('counterValue', JSON.stringify(minValue))
-        localStorage.setItem('counterMaxValue', JSON.stringify(maxValue))
-        localStorage.setItem('counterEndMinValue', JSON.stringify(endMinValue))
-        localStorage.setItem('counterEndMaxValue', JSON.stringify(endMaxValue))
-    }, [minValue, maxValue, endMinValue, endMaxValue])
+        if (isChecked) {
+            localStorage.setItem('counterMinValue', JSON.stringify(minValue))
+            localStorage.setItem('counterMaxValue', JSON.stringify(maxValue))
+        }
+        localStorage.setItem('counterInitValue', JSON.stringify(initValue))
+    }, [initValue])
 
     const incHandler = () => {
-        setMinValue(minValue + 1)
+        setInitValue(initValue + 1)
     }
     const resetHandler = () => {
-        localStorage.clear()
-        setMinValue(endMinValue)
+        if (!isChecked) {
+            localStorage.clear()
+            setInitValue(0)
+            localStorage.setItem('counterInitValue', JSON.stringify(0))
+        } else {
+            setInitValue(minValue ? minValue : 0)
+        }
+    }
+
+    const clearMinMaxStorage = () => {
+        if (isChecked) {
+            localStorage.removeItem('counterMinValue')
+            localStorage.removeItem('counterMaxValue')
+            setMaxValue(0)
+            setMinValue(0)
+        }
     }
 
     const setEditMode = () => {
         setChangeMode(false)
+        setInitValue(minValue)
     }
 
     const minOnChangeHandler = (e: ChangeEvent<HTMLInputElement>) => {
         let value = e.currentTarget.value
         setMinValue(Number(value))
-        setEndMinValue(Number(value))
-        if (Number(value) >= endMaxValue) {
+        setInitValue(0)
+        if (Number(value) >= maxValue) {
             setIncorrect(true)
         } else {
             setIncorrect(false)
@@ -68,8 +80,8 @@ const App = () => {
     const maxOnChangeHandler = (e: ChangeEvent<HTMLInputElement>) => {
         let value = e.currentTarget.value
         setMaxValue(Number(value))
-        setEndMaxValue(Number(value))
-        if (Number(value) <= endMinValue) {
+        setInitValue(0)
+        if (Number(value) <= minValue) {
             setIncorrect(true)
         } else {
             setIncorrect(false)
@@ -145,13 +157,19 @@ const App = () => {
             <div className="container">
                 <div className="value-container">
                     {incorrect
-                    ?   <p style={{color: "red"}}>Incorrect value!</p>
-                    :   <p style={minValue === maxValue ? {color: "red", fontSize: "32px", fontWeight: 900} : {}}>{!changeMode ? minValue : "enter values and press 'set'"}</p>
+                        ? <p style={{color: "red"}}>{incorrectValue}</p>
+                        : <p style={(initValue === maxValue && initValue !== 0) ? {
+                            color: "red",
+                            fontSize: "32px",
+                            fontWeight: 900
+                        } : {}}>{!changeMode ? initValue : "enter values and press 'set'"}</p>
                     }
                 </div>
                 <div className="button-container">
-                    <Button callback={incHandler} disabled={changeMode || (minValue === endMaxValue)} name={'inc'}/>
-                    <Button callback={resetHandler} disabled={changeMode || (minValue === endMinValue)} name={'reset'}/>
+                    <Button callback={incHandler} disabled={isChecked && (changeMode || (initValue === maxValue))}
+                            name={'inc'}/>
+                    <Button callback={resetHandler} disabled={isChecked && (changeMode || (initValue === minValue))}
+                            name={'reset'}/>
                 </div>
             </div>
         </div>
